+++ src/app/posts/[id]/page orig.jsx
@@ -1,6 +1,4 @@
 import { sql } from "@vercel/postgres";
-import Image from "next/image";
-import Link from "next/link";
 
 export default async function IndividualPost({ params }) {
   const result =
